Fix filename selection Dialog
d_rats/config.py:
  Use refactored d_rats/filenamebox.py.
  Designate when not a save file operation.

d_rats/filenamebox.py:
  Refactored from d_rats/miscwidgets.py.
  Fix GTK compatibility issues.
  Add save parameter to select proper action button.
  Add Unit test.

d_rats/miscwidgets.py:
  Remove FilenameBox class.

d_rats/qst.py:
  Use refactored d_rats/filenamebox.py.
  Fix GTK compatibility issues.

docs/source/d_rats.rst:
  Add d_rats/filenamebox.py
